Correction du type int de Product.price en float.

+++ product.py
@@ -1,8 +1,8 @@
 
 
 class Product:
-    def __init__(self, _name: str, _stock: int, _price: int, _is_unit: bool):
+    def __init__(self, _name: str, _stock: int, _price: float, _is_unit: bool):
         self.name: str = _name
         self.stock: int = _stock
-        self.price: int = _price
+        self.price: float = _price
         self.is_unit: bool = _is_unit
